衝突判定周りを改善
